+++ src/lib/WP1ContentListener.h
@@ -53,7 +53,7 @@ public:
 
 	void startDocument() { WPXContentListener::startDocument(); }
 	void startSubDocument() { WPXContentListener::startSubDocument(); }
-	void insertCharacter(uint16_t character);
+	void insertCharacter(uint32_t character);
 	void insertExtendedCharacter(uint8_t extendedCharacter);
 	void insertTab();
 	void insertBreak(uint8_t breakType) { WPXContentListener::insertBreak(breakType); }
