+++ src/lib/WP3ContentListener.h
@@ -55,7 +55,7 @@ public:
 
 	void startDocument() { WPXContentListener::startDocument(); };
 	void startSubDocument() { WPXContentListener::startSubDocument(); };
-	void insertCharacter(uint16_t character);
+	void insertCharacter(uint32_t character);
 	void insertTab();
 	void insertTab(uint8_t tabType, double tabPosition);
 	void insertBreak(uint8_t breakType) { WPXContentListener::insertBreak(breakType); };
