+++ src/lib/WP42ContentListener.h
@@ -48,7 +48,7 @@ public:
 
 	void startDocument() { WPXContentListener::startDocument(); }
 	void startSubDocument() { WPXContentListener::startSubDocument(); }
-	void insertCharacter(uint16_t character);
+	void insertCharacter(uint32_t character);
 	void insertTab(uint8_t tabType, double tabPosition);
 	void insertBreak(uint8_t breakType) { WPXContentListener::insertBreak(breakType); }
 	void insertEOL();
