+++ src/lib/WP5Parser.cpp
@@ -94,14 +94,14 @@ void WP5Parser::parseDocument(WPXInputStream *input, WPXEncryption *encryption, 
 					listener->insertEOL();
 					break;
 				case 0x0B: // soft new page (convert like space)
-					listener->insertCharacter((uint16_t) ' ');
+					listener->insertCharacter((uint32_t) ' ');
 					listener->insertBreak(WPX_SOFT_PAGE_BREAK);
 					break;
 				case 0x0C: // hard new page
 					listener->insertBreak(WPX_PAGE_BREAK);
 					break;
 				case 0x0D: // soft new line (convert like space)
-					listener->insertCharacter((uint16_t) ' ');
+					listener->insertCharacter((uint32_t) ' ');
 					break;
 				default:
 					// unsupported or undocumented token, ignore
